Use French region names when finding French subdivisions
If a French region name is found for a text related to the France country,
then the department of the regional prefecture is returned as the
matching subdivision.

This works with old (pre-2016) and new (post-2016) region names, with a
special case for the old region name "Centre", which is not used to
avoid false postives (with, e.g., "Centre Pompidou"). To match this
old region name, there must be "region centre" or "conseil regional
centre" in the input text.